requirements.txt
APScheduler==3.6.3
boto3==1.18.19
botocore==1.21.19
cachetools==4.2.1
certifi==2020.6.20
chardet==4.0.0
greenlet==1.1.2
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
packaging==20.9
prompt-toolkit==3.0.16
protobuf==3.15.4
psycopg2==2.9.3
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.15.0
tmdbsimple==2.8.0
tornado==6.1
tweepy==3.10.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
wcwidth==0.2.5
wit==6.0.0
spacy==3.0.6
SQLAlchemy==1.4.35
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
